LayoutTests/imported/w3c/web-platform-tests/css/selectors/is-specificity-expected.txt


FAIL Test :is() uses highest possible specificity assert_equals: expected "20px" but got "10px"
